requirements.txt
flask
dash
pandas
numpy
matplotlib
mysql-connector-python
kafka-python
plotly
gunicorn
